Make SingleChildControlCollection public
